Estimate Amino Acid Usage Frequencies of CDSs.

Estimate Amino Acid Usage Frequencies of CDSs.

Usage

est_aau(cf, codon_table = get_codon_table())

Arguments

cf

matrix of codon frequencies as calculated by `count_codons()`.

codon_table

codon_table a table of genetic code derived from get_codon_table or create_codon_table.

Value

a data.table with amino acid frequencies of CDSs. The columns include three-letter abbreviation of the amino acid, single-letter abbreviation, usage frequency of the amino acid in all sequences, and usage frequency proportion.

Examples

# estimate amino acid frequencies of yeast genes
cf_all <- count_codons(yeast_cds)
aau <- est_aau(cf_all)
print(aau)
#>     amino_acid aa_code  count proportion
#>         <char>  <char>  <num>      <num>
#>  1:        Phe       F 133094 0.04567496
#>  2:        Leu       L 281010 0.09643651
#>  3:        Ser       S 263057 0.09027543
#>  4:        Tyr       Y  97696 0.03352714
#>  5:        Cys       C  38839 0.01332870
#>  6:        Trp       W  30566 0.01048958
#>  7:        Pro       P 125783 0.04316598
#>  8:        His       H  62347 0.02139613
#>  9:        Gln       Q 113195 0.03884606
#> 10:        Arg       R 130244 0.04469690
#> 11:        Ile       I 191381 0.06567779
#> 12:        Met       M  61057 0.02095343
#> 13:        Thr       T 170219 0.05841545
#> 14:        Asn       N 176635 0.06061728
#> 15:        Lys       K 212102 0.07278878
#> 16:        Val       V 163969 0.05627059
#> 17:        Ala       A 160384 0.05504029
#> 18:        Asp       D 167835 0.05759731
#> 19:        Glu       E 188978 0.06485313
#> 20:        Gly       G 145547 0.04994856
#>     amino_acid aa_code  count proportion