get tRNA gene copy number from GtRNADB

extract_trna_gcn get tRNA gene copy number from GtRNADB

Usage

extract_trna_gcn(trna_seq)

Arguments

trna_seq

a fasta file of tRNA sequences from GtRNADB

Value

a table of tRNA gene copy number for each anticodon. The name of the element corresponds to the anticodon and the amino acid it carries. tRNA transporting either initiator methionine (iMet) in eukaryotes and archaea, or formylmethionine (fMet) in prokaryotes, utilized to initiate translation, is excluded from the count.

Examples

# get tRNA gene copy number for yeast
trna_gcn <- extract_trna_gcn(yeast_trna)