Amino Acid Usage

Calculate Amino Acid Usage Frequencies of each CDS.

Usage

get_aau(cf, codon_table = get_codon_table())

Arguments

cf

matrix of codon frequencies as calculated by `count_codons()`.

codon_table

a table of genetic code derived from get_codon_table or create_codon_table.

Value

a matrix of amino acid frequencies for each CDS. Each row corresponds to a sequence, and each column represents an amino acid.

Examples

# estimate amino acid frequencies of yeast CDSs
cf_all <- count_codons(yeast_cds)
aau_gene <- get_aau(cf_all)
head(aau_gene)
#>                Ala        Arg        Asn        Asp         Cys         Gln
#> YPL071C 0.04487179 0.06410256 0.08974359 0.12820513 0.006410256 0.012820513
#> YLL050C 0.07692308 0.04895105 0.04195804 0.08391608 0.006993007 0.006993007
#> YMR172W 0.05563282 0.05006954 0.11821975 0.05841446 0.002781641 0.050069541
#> YOR185C 0.07727273 0.03636364 0.05454545 0.06363636 0.013636364 0.050000000
#> YLL032C 0.04242424 0.03515152 0.08484848 0.04242424 0.010909091 0.046060606
#> YBR225W 0.05444444 0.04333333 0.06888889 0.06444444 0.008888889 0.031111111
#>                Glu        Gly         His        Ile        Leu        Lys
#> YPL071C 0.04487179 0.03846154 0.032051282 0.08974359 0.06410256 0.09615385
#> YLL050C 0.06993007 0.06293706 0.006993007 0.03496503 0.07692308 0.07692308
#> YMR172W 0.04728790 0.05285118 0.022253129 0.06119611 0.08484006 0.05702364
#> YOR185C 0.06363636 0.05909091 0.022727273 0.05000000 0.07727273 0.08181818
#> YLL032C 0.06666667 0.03878788 0.019393939 0.10060606 0.09090909 0.07151515
#> YBR225W 0.05444444 0.03777778 0.031111111 0.06111111 0.08555556 0.08000000
#>                Met        Phe        Pro        Ser        Thr         Trp
#> YPL071C 0.03205128 0.02564103 0.02564103 0.05769231 0.03846154 0.032051282
#> YLL050C 0.01398601 0.05594406 0.02797203 0.11188811 0.05594406 0.006993007
#> YMR172W 0.02642559 0.02086231 0.04311544 0.11265647 0.08344924 0.004172462
#> YOR185C 0.01818182 0.05454545 0.05454545 0.02727273 0.05909091 0.013636364
#> YLL032C 0.02545455 0.05454545 0.04000000 0.08121212 0.05939394 0.002424242
#> YBR225W 0.01444444 0.04444444 0.05888889 0.13444444 0.05444444 0.011111111
#>                Tyr        Val
#> YPL071C 0.03846154 0.03846154
#> YLL050C 0.04895105 0.08391608
#> YMR172W 0.01808067 0.03059805
#> YOR185C 0.04090909 0.08181818
#> YLL032C 0.04363636 0.04363636
#> YBR225W 0.02888889 0.03222222