yeast mRNA expression levels

Yeast mRNA FPKM determined from rRNA-depleted (RiboZero) total RNA-Seq libraries. RUN1_0_WT and RUN2_0_WT (0 min after RNA Pol II repression) were averaged and used here.

Usage

yeast_exp

Format

a data.frame with 6717 rows and three columns:

gene_id

gene ID

gene_name

gene name

fpkm

mRNA expression level in Fragments per kilobase per million reads

Source

<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE57385>

References

Presnyak V, Alhusaini N, Chen YH, Martin S, Morris N, Kline N, Olson S, Weinberg D, Baker KE, Graveley BR, et al. 2015. Codon optimality is a major determinant of mRNA stability. Cell 160:1111-1124.

Examples

head(yeast_exp)
#> # A tibble: 6 × 3
#>   gene_id gene_name  fpkm
#>   <chr>   <chr>     <dbl>
#> 1 YAL001C TFC3       28.8
#> 2 YAL002W VPS8       19.2
#> 3 YAL003W EFB1      889. 
#> 4 YAL004W YAL004w    84.1
#> 5 YAL005C SSA1      385. 
#> 6 YAL007C ERP2      123.